pyproject.toml
[tool.poetry]
name = "dotobot"
version = "0.1.0"
description = "A discord bot to keep track of out-of-context quotes and mod abuse"
authors = ["Sam van der Kris <samvdkris@disroot.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
